pyproject.toml
[tool.poetry]
name = "kfsearch"
version = "0.1.0"
description = ""
authors = ["Andy in Berlin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.1"
plotly = "^5.24.1"
dash = "^2.18.1"
speechrecognition = "^3.10.4"
pydub = "^0.25.1"
soundfile = "^0.12.1"
torch = "^2.4.1"
whisper = "^1.1.10"
pocketsphinx = "^5.0.3"
google-cloud-speech = "^2.27.0"
python-dotenv = "^1.0.1"
typer = "^0.12.5"
bs4 = "^0.0.2"
nbconvert = "^7.16.4"
flask = "^3.0.3"
dash-bootstrap-components = "^1.6.0"
pandas = "^2.2.3"
elasticsearch = "^8.15.1"
jupyter = "^1.1.1"
black = "^24.10.0"
google-cloud-storage = "^2.18.2"
flake8 = "^7.1.1"
loguru = "^0.7.3"
pytest = "^8.3.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
